renamed to dbai and integrated with eval sql_generator
